#307 Move upload test to test_client
